[Refactor] #127 - 홈 MVI 통합 구현
